+++ src/Command/SubmitCommand.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace IndexNowKit\SymfonyBundle\Command;
 
-use IndexNowKit\IndexNowKit;
+use IndexNowKit\Console\SubmitRunner;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
@@ -16,7 +16,7 @@ use Symfony\Component\Console\Style\SymfonyStyle;
 #[AsCommand(name: 'indexnow:submit', description: 'Submit URLs to IndexNow immediately (synchronously, bypassing the queue)')]
 final class SubmitCommand extends Command
 {
-    public function __construct(private readonly IndexNowKit $indexNow, private readonly SubmitterFactoryInterface $submitters, private readonly ResultFormatterInterface $formatter = new ResultRenderer())
+    public function __construct(private readonly SubmitRunner $runner)
     {
         parent::__construct();
     }
@@ -32,13 +32,9 @@ final class SubmitCommand extends Command
 
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
-        $io = new SymfonyStyle($input, $output);
         /** @var list<string> $urls */
         $urls = $input->getArgument('urls');
-        $force = (bool) $input->getOption('force');
-        $dryRun = (bool) $input->getOption('dry-run');
-        $submitter = $force || $dryRun ? $this->submitters->create($force, $dryRun) : $this->indexNow->submitter;
 
-        return $this->formatter->results($io, $submitter->submit($urls), (bool) $input->getOption('json'));
+        return $this->runner->run(new SymfonyStyle($input, $output), $urls, (bool) $input->getOption('force'), (bool) $input->getOption('dry-run'), (bool) $input->getOption('json'));
     }
 }